+++ js/chat.html
@@ -4,10 +4,12 @@
     <script src="./sock.js"></script>
    
     <script src="./channelsPanel.js"></script>
-    <script src="chat.js"></script>
     <script type="module" src="config.js"></script>
     <script type="module" src="login.js"></script>
-    
+    <script type="module" src="chat.js"></script>
+    <script type="module" src="elements.js"></script>
+    <script type="module" src="eventBus.js"></script>   
+    <script type="module" src="main.js"></script> 
     <link rel="stylesheet" type="text/css" href="specific.css" ></script>
     <link rel="stylesheet" type="text/css" href="general.css" ></script>
